+++ packages/components/tooltip/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @mochi-ui/tooltip
 
+## 0.2.20
+
+### Patch Changes
+
+- Updated dependencies [[`e0cf264`](https://github.com/consolelabs/mochi-ui/commit/e0cf26419ff16ebd3ffcd577ec44896ab2c7e836)]:
+  - @mochi-ui/theme@0.20.0
+
+## 0.2.19
+
+### Patch Changes
+
+- Updated dependencies [[`c393607`](https://github.com/consolelabs/mochi-ui/commit/c39360797bb96792c2c82287e79bcdf9eddf3641)]:
+  - @mochi-ui/theme@0.19.0
+
 ## 0.2.18
 
 ### Patch Changes